Code Refactor

+++ src/main/java/com/lap/lapproject/model/Trainer.java
@@ -19,10 +19,7 @@ public class Trainer extends User {
     public Trainer() {
     }
 
-
-
-
-    public Trainer(long id, String fName, String lName, String email, String phoneNmbr, Boolean activeStatus) {
+    public Trainer(int id, String fName, String lName, String email, String phoneNmbr, Boolean activeStatus) {
         super(id, fName, lName, email, phoneNmbr, activeStatus);
     }
 }
